Added fixes and missing endpoints
